@@ -1,17 +1,13 @@
 import adapter from "@sveltejs/adapter-vercel";
 import { vitePreprocess } from "@sveltejs/vite-plugin-svelte";
 import { mdsvex } from "mdsvex";
+import mdsvexConfig from "./mdsvex.config.js";
 
 /** @type {import('@sveltejs/kit').Config} */
 const config = {
   // Consult https://kit.svelte.dev/docs/integrations#preprocessors
   // for more information about preprocessors
-  preprocess: [
-    vitePreprocess(),
-    mdsvex({
-      extensions: [".md"],
-    }),
-  ],
+  preprocess: [vitePreprocess(), mdsvex(mdsvexConfig)],
 
   extensions: [".svelte", ".md"],
 
@@ -22,6 +18,10 @@ const config = {
     adapter: adapter(),
     paths: {
       base: "",
+    },
+    alias: {
+      $posts: "./src/posts/",
+      $static: "./static/",
     },
   },
 };
